refactor(web): Phase 6 — lift tasks & sessions into 3-tier (Infra → Service → API)
Issue #317. Most complex phase of the 3-tier migration. Absorbs the legacy
`lib/{task-runner,task-store,pending-inputs,session-store,agent-pool}.ts`
into the new tier layout, and migrates the `tasks.*` / `sessions.*` tRPC
sub-routers from the legacy router into dedicated `server/api/*/router.ts`
files.

- Infra:
  - `server/infra/db/queries/tasks.ts` — `TaskQueries` class + prune
    scheduler, replacing `lib/task-store.ts`.
  - `server/infra/agents/agent-pool.ts` — moved from `lib/agent-pool.ts`.
- Services:
  - `server/services/task-service.ts` — absorbs `task-runner` (submit /
    abort / cancel + agent event loop), `pending-inputs`
    (AskUserQuestion / ExitPlanMode resolver map), and `session-store`
    (event-buffer read helpers). Introduces typed `TaskConflictError` /
    `WorkspaceNotFoundError` for the API tier.
  - `server/services/session-service.ts` — thin façade over
    agent.listSessions().
- API:
  - `server/api/tasks/router.ts` — `list`, `submit`, `get`, `isRunning`,
    `abort`, `cancel`, `rerun`. Translates service errors to TRPCError
    codes.
  - `server/api/sessions/router.ts` — `list`. Same shape.
  - Merged into `appRouter` via `t.mergeRouters(legacy, …)`; the legacy
    `tasksRouter` / `sessionsRouter` are removed and the migration
    invariant comment is updated.

Behaviour-preserving: the persisted-task schema is unchanged, the
`globalThis` symbol keys (`band.task-runner.*`, `band.pending-inputs`)
keep their legacy strings so singletons stay consistent across module
reloads, and the ESM cycle between `workspace-service` ↔ `lib/workspace`
↔ `task-service` is documented at every leg.
