+++ apps/web/src/server/services/workspace-service.ts
@@ -21,22 +21,22 @@ import {
   type WorktreeState,
   worktreesDir,
 } from "../../lib/state";
-// FRAGILE: ESM cycle leg — `lib/task-runner` imports `lib/workspace`,
+import { emit } from "../../lib/watcher";
+// FRAGILE: ESM cycle leg — `services/task-service` imports `lib/workspace`,
 // which imports `workspaceService` from this file. The cycle is safe only
 // because every `workspaceService` reference is inside a function body
 // (live binding). See `lib/workspace.ts` for the cycle note before
 // capturing `submitTask` (or anything else here) at module load.
-import { submitTask } from "../../lib/task-runner";
-import { deleteWorkspaceTasks } from "../../lib/task-store";
-import { emit } from "../../lib/watcher";
+import { TaskQueries } from "../infra/db/queries/tasks";
 import { WorkspaceQueries } from "../infra/db/queries/workspaces";
 // FRAGILE: ESM cycle leg #2 — `./cronjob-service` imports `submitTask`
-// from `lib/task-runner`, which imports `lib/workspace`, which imports
+// from `./task-service`, which imports `lib/workspace`, which imports
 // `workspaceService` from this file. Same live-binding constraint as the
 // `submitTask` import above: keep every `cronjobService` reference inside
 // a function body. Capturing `const cs = cronjobService;` at module load
 // on this leg would silently get `undefined`.
 import { cronjobService } from "./cronjob-service";
+import { submitTask } from "./task-service";
 import { terminalService } from "./terminal-service";
 
 const execFileAsync = promisify(execFile);
@@ -419,7 +419,7 @@ export class WorkspaceService {
     // suppress the `emit` below, otherwise the dashboard would keep
     // showing the just-deleted workspace.
     try {
-      const deletedTasks = deleteWorkspaceTasks(workspaceId);
+      const deletedTasks = new TaskQueries().deleteWorkspaceTasks(workspaceId);
       if (deletedTasks > 0) {
         log.info({ workspaceId, count: deletedTasks }, "deleted workspace tasks on removal");
       }
