+++ apps/web/tests/task-cleanup.test.ts
@@ -19,7 +19,7 @@ import { DatabaseSync } from "node:sqlite";
 import { drizzle } from "drizzle-orm/node-sqlite";
 import { migrate } from "drizzle-orm/node-sqlite/migrator";
 import { afterAll, beforeAll, describe, expect, it } from "vitest";
-import { TASK_RETENTION_MS } from "../src/lib/task-store";
+import { TASK_RETENTION_MS } from "../src/server/infra/db/queries/tasks";
 import { seedSettings, seedState } from "./helpers/seed-state";
 import { SERVER_RUNTIME, SERVER_SCRIPT } from "./helpers/server-runtime";
